requirement.txt
wheel
tqdm
pytorch_lightning
black
ruff
hydra-core
jaxtyping
beartype
wandb
einops
colorama
scikit-image
colorspacious
matplotlib
moviepy
imageio
git+https://github.com/dcharatan/diff-gaussian-rasterization-modified
timm
dacite
lpips
e3nn
plyfile
tabulate
svg.py
opencv-python==4.6.0.66
sk-video
deformable_aggregation_ext==0.0.0